Add files via upload
The code that is provided to solve the problem falls under Java programming language,
resides in the Project1 class with AttendanceSystem being a static nested class. This
system is designed to interactively collect and manage attendance data for users
categorized into three roles: This is a culmination of the Admin, Faculty, and Student.
Another aspect is the user interface that the console provides directions to the user to
enter name, choose the role from a specified list and enter the ID number.
When run, the main method of the application creates an instance of the
AttendanceSystem class and calls the run method, which starts the loop of user
interactions. This make the users to enter his or her name and role repeatedly before
choosing to log out of the system. When the entered role is correct (Admin, Faculty, or
Student), the system prompts the user to enter the ID number and then saves all acquired
data in three `ArrayList`s for names, positions, and IDs.
The features of the program involve being able to save the gathered information
regarding attendance to a text file under the name of ‘NewProject. txt’ when the
program is closed. The writeFile method is responsible for writing the said user’s name,
role, and their ID number adjacent to one another onto the output file, along with
standard messages that reflect successful attendance marking. This helps in retaining
all the session data for future use ad last longer than the normal session data.
Also, the code suggests that the application relies on an external class pro2 from which
methods thank and thank1 are invoked to likely show the users thank yous. This aspect
though, is not elaborated to the extent within the given snippet of the paper.
All in all, this attendance system shows the basic concepts of programming like
handling inputs from the users, collection data type, and file handling I/O which makes
this a good example for learning or even for little organization’s attendance tracking.